requirements.txt
django
MySQL-python
django-rosetta
djangoembed
sorl-thumbnail
django-filebrowser
feedparser
simplejson
httplib2
PIL
django-debug-toolbar
django-annoying
beautifulsoup
flup
python-memcached
django-admin-tools

